RavenDB-14706 Move RelationalDatabase directory to the ETL > Providers
