+++ src/Raven.Server/Documents/ETL/Providers/RelationalDatabase/RelationalWriters/RelationalDatabaseWriteStats.cs
@@ -1,4 +1,4 @@
-﻿namespace Raven.Server.Documents.ETL.Relational.RelationalWriters;
+﻿namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase.RelationalWriters;
 
 public sealed class RelationalDatabaseWriteStats
 {
