+++ src/Raven.Server/Documents/ETL/Providers/RelationalDatabase/Test/RelationalDatabaseEtlTestScriptResult.cs
@@ -2,7 +2,7 @@
 using Raven.Server.Documents.ETL.Test;
 using Raven.Server.NotificationCenter.Notifications.Details;
 
-namespace Raven.Server.Documents.ETL.Relational.Test;
+namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase.Test;
 
 public sealed class RelationalDatabaseEtlTestScriptResult : TestEtlScriptResult
 {
