+++ src/Raven.Server/Documents/ETL/Providers/RelationalDatabase/TestRelationalDatabaseEtlScript.cs
@@ -1,0 +1,14 @@
+using Raven.Client.Documents.Operations.ConnectionStrings;
+using Raven.Client.Documents.Operations.ETL;
+using Raven.Server.Documents.ETL.Test;
+
+namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+
+public sealed class TestRelationalDatabaseEtlScript<TRelationalConnectionString, TRelationalEtlConfiguration> : TestEtlScript<TRelationalEtlConfiguration, TRelationalConnectionString>
+    where TRelationalConnectionString: ConnectionString
+    where TRelationalEtlConfiguration: EtlConfiguration<TRelationalConnectionString>
+{
+    public bool PerformRolledBackTransaction;
+
+    public TRelationalConnectionString Connection;
+}
