+++ src/Raven.Server/Documents/ETL/Providers/SQL/Handlers/Processors/SqlEtlHandlerProcessorForTest.cs
@@ -1,8 +1,7 @@
 ﻿using JetBrains.Annotations;
 using Raven.Client.Documents.Operations.ETL.SQL;
 using Raven.Server.Documents.ETL.Providers.Raven.Handlers.Processors;
-using Raven.Server.Documents.ETL.Providers.SQL.RelationalWriters;
-using Raven.Server.Documents.ETL.Relational;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
 using Raven.Server.Json;
 using Sparrow.Json;
