+++ src/Raven.Server/Documents/ETL/Providers/Snowflake/SnowflakeDocumentTransformer.cs
@@ -2,20 +2,21 @@
 using System.Linq;
 using Raven.Client.Documents.Operations.ETL;
 using Raven.Client.Documents.Operations.ETL.Snowflake;
-using Raven.Server.Documents.ETL.Relational;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
 using Raven.Server.Documents.Patch;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 
 namespace Raven.Server.Documents.ETL.Providers.Snowflake;
 
-internal sealed class SnowflakeDocumentTransformer : RelationalDatabaseDocumentTransformerBase<SnowflakeConnectionString, SnowflakeEtlConfiguration>
+internal sealed class SnowflakeDocumentTransformer(
+    Transformation transformation,
+    DocumentDatabase database,
+    DocumentsOperationContext context,
+    SnowflakeEtlConfiguration config)
+    : RelationalDatabaseDocumentTransformerBase<SnowflakeConnectionString, SnowflakeEtlConfiguration>(transformation, database, context, config,
+        PatchRequestType.SnowflakeEtl)
 {
-    public SnowflakeDocumentTransformer(Transformation transformation, DocumentDatabase database, DocumentsOperationContext context, SnowflakeEtlConfiguration config) :
-        base(transformation, database, context, config, PatchRequestType.SnowflakeEtl)
-    {
-    }
-    
     protected override List<RelationalDatabaseTableWithRecords> GetEtlTables()
     {
         return _config.SnowflakeTables.Select(RelationalDatabaseTableWithRecords.FromSnowflakeEtlTable).ToList();
