+++ src/Raven.Server/Documents/ETL/Providers/Snowflake/SnowflakeEtl.cs
@@ -1,8 +1,8 @@
 ﻿using Raven.Client.Documents.Operations.ETL;
 using Raven.Client.Documents.Operations.ETL.Snowflake;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.RelationalWriters;
 using Raven.Server.Documents.ETL.Providers.Snowflake.RelationalWriters;
-using Raven.Server.Documents.ETL.Relational;
-using Raven.Server.Documents.ETL.Relational.RelationalWriters;
 using Raven.Server.Documents.ETL.Stats;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
