+++ test/SlowTests/Server/Documents/ETL/SQL/RavenDB_19518.cs
@@ -15,8 +15,8 @@ using System.Threading;
 using System;
 using Tests.Infrastructure;
 using Npgsql;
-using Raven.Server.Documents.ETL.Relational;
-using Raven.Server.Documents.ETL.Relational.Test;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Test;
 
 namespace SlowTests.Server.Documents.ETL.SQL;
