+++ test/SlowTests/Server/Documents/ETL/SQL/SqlEtlTests.cs
@@ -19,14 +19,12 @@ using Raven.Client.Documents;
 using Raven.Client.Documents.Operations.Attachments;
 using Raven.Client.Documents.Operations.ConnectionStrings;
 using Raven.Client.Documents.Operations.ETL;
-using Raven.Client.Documents.Operations.ETL.Snowflake;
 using Raven.Client.Documents.Operations.ETL.SQL;
 using Raven.Client.Extensions;
 using Raven.Server.Config;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Test;
 using Raven.Server.Documents.ETL.Providers.SQL;
-using Raven.Server.Documents.ETL.Providers.SQL.RelationalWriters;
-using Raven.Server.Documents.ETL.Relational;
-using Raven.Server.Documents.ETL.Relational.Test;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.SqlMigration;
 using Raven.Tests.Core.Utils.Entities;
@@ -36,19 +34,14 @@ using Tests.Infrastructure;
 using Tests.Infrastructure.ConnectionString;
 using Xunit;
 using Xunit.Abstractions;
-using SnowflakeConnectionString = Raven.Client.Documents.Operations.ETL.Snowflake.SnowflakeConnectionString;
 
 namespace SlowTests.Server.Documents.ETL.SQL
 {
-    public class SqlEtlTests : RavenTestBase
+    public class SqlEtlTests(ITestOutputHelper output) : RavenTestBase(output)
     {
-        public SqlEtlTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
         private readonly List<string> _dbNames = new List<string>();
 
-        protected const string defaultScript = @"
+        private const string DefaultScript = @"
 var orderData = {
     Id: id(this),
     OrderLinesCount: this.OrderLines.length,
@@ -96,7 +89,7 @@ loadToOrders(orderData);
 
                     var etlDone = Etl.WaitForEtlToComplete(store, (n, s) => GetOrdersCount(connectionString) == testCount);
 
-                    SetupSqlEtl(store, connectionString, defaultScript);
+                    SetupSqlEtl(store, connectionString, DefaultScript);
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
 
@@ -185,7 +178,7 @@ DROP DATABASE [SqlReplication-{dbName}]";
 
                     var etlDone = Etl.WaitForEtlToComplete(store);
 
-                    SetupSqlEtl(store, connectionString, defaultScript);
+                    SetupSqlEtl(store, connectionString, DefaultScript);
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
 
@@ -448,7 +441,7 @@ loadToOrders(orderData);");
 
                     var etlDone = Etl.WaitForEtlToComplete(store);
 
-                    SetupSqlEtl(store, connectionString, "if(this.OrderLines.length > 0) { \r\n" + defaultScript + " \r\n}");
+                    SetupSqlEtl(store, connectionString, "if(this.OrderLines.length > 0) { \r\n" + DefaultScript + " \r\n}");
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
 
@@ -508,7 +501,7 @@ loadToOrders(orderData);");
 
                     var etlDone = Etl.WaitForEtlToComplete(store);
 
-                    SetupSqlEtl(store, connectionString, defaultScript);
+                    SetupSqlEtl(store, connectionString, DefaultScript);
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
 
@@ -554,7 +547,7 @@ loadToOrders(orderData);");
 
                     var etlDone = Etl.WaitForEtlToComplete(store);
 
-                    SetupSqlEtl(store, connectionString, defaultScript);
+                    SetupSqlEtl(store, connectionString, DefaultScript);
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
 
@@ -597,7 +590,7 @@ loadToOrders(orderData);");
 
                     var etlDone = Etl.WaitForEtlToComplete(store);
 
-                    SetupSqlEtl(store, connectionString, defaultScript, insertOnly: true);
+                    SetupSqlEtl(store, connectionString, DefaultScript, insertOnly: true);
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
 
@@ -748,7 +741,7 @@ var nameArr = this.StepName.split('.'); loadToOrders({});");
                                     {
                                         new Transformation()
                                         {
-                                            Collections = { "Orders" }, Name = "OrdersAndLines", Script = defaultScript + "output('test output')"
+                                            Collections = { "Orders" }, Name = "OrdersAndLines", Script = DefaultScript + "output('test output')"
                                         }
                                     }
                                 }
@@ -829,7 +822,7 @@ var nameArr = this.StepName.split('.'); loadToOrders({});");
                                         new SqlEtlTable { TableName = "OrderLines", DocumentIdColumn = "OrderId" },
                                         new SqlEtlTable { TableName = "NotUsedInScript", DocumentIdColumn = "OrderId" },
                                     },
-                                    Transforms = { new Transformation() { Collections = { "Orders" }, Name = "OrdersAndLines", Script = defaultScript } }
+                                    Transforms = { new Transformation() { Collections = { "Orders" }, Name = "OrdersAndLines", Script = DefaultScript } }
                                 }
                             }, database, database.ServerStore, context);
                         
@@ -1165,7 +1158,7 @@ loadToOrders(orderData);
 
                     var etlDone = Etl.WaitForEtlToComplete(store, numOfProcessesToWaitFor: 2);
 
-                    SetupSqlEtl(store, connectionString, defaultScript, collections: new List<string> { "Orders", "FavouriteOrders" });
+                    SetupSqlEtl(store, connectionString, DefaultScript, collections: new List<string> { "Orders", "FavouriteOrders" });
 
                     etlDone.Wait(TimeSpan.FromMinutes(5));
